requirements.txt
# Requirements para desenvolvimento local
# Instale com: pip install -r requirements.txt

# Clientes de acesso aos dados
boto3>=1.28.0              # Cliente S3/MinIO
trino>=0.328.0             # Cliente Trino
pandas>=2.0.0              # Manipulação de dados
pyarrow>=12.0.0            # Suporte a Parquet
requests>=2.31.0           # HTTP requests

# PySpark (para desenvolvimento local)
pyspark>=3.5.0             # Apache Spark

# Airflow providers (para desenvolvimento de DAGs)
apache-airflow>=2.7.0
apache-airflow-providers-apache-spark>=4.1.0

# Jupyter (opcional, para notebooks)
jupyter>=1.0.0
ipython>=8.0.0

# Visualização
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Qualidade de código
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
